Dev/4.5.0

+++ Android/APIExample/cloud_build.sh
@@ -12,27 +12,34 @@ sed -i -e "s#YOUR APP ID#${APP_ID}#g" app/src/main/res/values/string_configs.xml
 sed -i -e "s#YOUR APP CERTIFICATE##g" app/src/main/res/values/string_configs.xml
 sed -i -e "s#YOUR ACCESS TOKEN##g" app/src/main/res/values/string_configs.xml
 rm -f app/src/main/res/values/string_configs.xml-e
+echo "First argument: $1"
+echo "Second argument: $2"
+if [ "$1" = "false" ]; then
+     echo "clould build sdk_url is empty"
+else
+    echo "clould build sdk_url is not empty"
+    ## config simple filter
+    sed -i -e "s#simpleFilter = false#simpleFilter = true#g" gradle.properties
+    rm -f gradle.properties-e
+    mkdir -p agora-simple-filter/src/main/agoraLibs
+    cp -r ../../sdk/arm64-v8a agora-simple-filter/src/main/agoraLibs/
+    cp -r ../../sdk/armeabi-v7a agora-simple-filter/src/main/agoraLibs/
+    curl -o opencv4.zip https://agora-adc-artifacts.s3.cn-north-1.amazonaws.com.cn/androidLibs/opencv4.zip
+    unzip opencv4.zip
+    mkdir -p agora-simple-filter/src/main/libs
+    mv arm64-v8a agora-simple-filter/src/main/libs
+    mv armeabi-v7a agora-simple-filter/src/main/libs
+    sed -i -e "s#jniLibs/#libs/#g" agora-simple-filter/src/main/cpp/CMakeLists.txt
+    rm -f agora-simple-filter/src/main/cpp/CMakeLists.txt-e
 
-## config simple filter
-sed -i -e "s#simpleFilter = false#simpleFilter = true#g" gradle.properties
-rm -f gradle.properties-e
-mkdir -p agora-simple-filter/src/main/agoraLibs
-cp -r ../../sdk/arm64-v8a agora-simple-filter/src/main/agoraLibs/
-cp -r ../../sdk/armeabi-v7a agora-simple-filter/src/main/agoraLibs/
-curl -o opencv4.zip https://agora-adc-artifacts.s3.cn-north-1.amazonaws.com.cn/androidLibs/opencv4.zip
-unzip opencv4.zip
-mkdir -p agora-simple-filter/src/main/libs
-mv arm64-v8a agora-simple-filter/src/main/libs
-mv armeabi-v7a agora-simple-filter/src/main/libs
-sed -i -e "s#jniLibs/#libs/#g" agora-simple-filter/src/main/cpp/CMakeLists.txt
-rm -f agora-simple-filter/src/main/cpp/CMakeLists.txt-e
+    ## config agora stream encrypt
+    sed -i -e "s#streamEncrypt = false#streamEncrypt = true#g" gradle.properties
+    rm -f gradle.properties-e
+    mkdir -p agora-stream-encrypt/src/main/agoraLibs
+    cp -r ../../sdk/arm64-v8a agora-stream-encrypt/src/main/agoraLibs/
+    cp -r ../../sdk/armeabi-v7a agora-stream-encrypt/src/main/agoraLibs/
+fi
 
-## config agora stream encrypt
-sed -i -e "s#streamEncrypt = false#streamEncrypt = true#g" gradle.properties
-rm -f gradle.properties-e
-mkdir -p agora-stream-encrypt/src/main/agoraLibs
-cp -r ../../sdk/arm64-v8a agora-stream-encrypt/src/main/agoraLibs/
-cp -r ../../sdk/armeabi-v7a agora-stream-encrypt/src/main/agoraLibs/
 
 ## config beauty
 sed -i -e "s#io.agora.api.example#io.agora.entfull#g" app/build.gradle
@@ -52,7 +59,23 @@ cd - || exit 1
 ./gradlew clean || exit 1
 ./gradlew :app:assembleRelease || exit 1
 
+SDK_VERSION=""
+if [ "$1" = "false" ]; then
+    sdk_version_file="./gradle.properties"
+    if [[ -f "$sdk_version_file" ]]; then
+    rtc_sdk_version=$(grep "rtc_sdk_version" "$sdk_version_file" | cut -d'=' -f2)
+    if [[ -n "$rtc_sdk_version" ]]; then
+        SDK_VERSION=$(echo "$rtc_sdk_version" | sed 's/^[ \t]*//;s/[ \t]*$//')
+    else
+        echo "rtc_sdk_version value not found"
+    fi
+else
+    echo "file not found: $sdk_version_file"
+fi
+else
+    SDK_VERSION=$(echo $sdk_url | cut -d "/" -f 5)
+fi
+
 if [ "$WORKSPACE" != "" ]; then
-SDK_VERSION=$(echo $sdk_url | cut -d "/" -f 5)
 cp app/build/outputs/apk/release/*.apk $WORKSPACE/APIExample_${BUILD_NUMBER}_${SDK_VERSION}_$(date "+%Y%m%d%H%M%S").apk
 fi
